Remove package pkg/netutils
